Update SROC fixtures used
We realised that we could use the 'previous' fixture both for our new supplementary test and the annual tests. This saves a bunch of duplication in the integration sets. It still creates a bill run but it is for the previous year. If anything it's a more robust way of checking that bill runs in previous billing periods shouldn't prevent a new one from being created.