Update tests due to supplementary year bump (#86)
https://eaflood.atlassian.net/browse/WATER-4403

In [Bump supplementary end year if no annual bill run](https://github.com/DEFRA/water-abstraction-system/pull/875) we amended our billing engine for supplementary. We made it possible for users to generate supplementary bill runs in years where no annual has yet been created and sent.

Previously, the Billing & Data team would have to put a block on creating bill runs until the annuals were generated. This is because supplementary takes into account previous transactions whereas annual doesn't. Without the change creating a supplementary followed by an annual bill run would result in the customer getting charged twice.

We now determine when the last annual bill run was sent and use its financial end year as the end year for the supplementary bill run. Annual created in this billing period? Then the supplementary will start there and work back 5 years. If the last sent annual was for the previous billing period the supplementary will start _there_ and work back 5 years.

The problem is this means our billing engine is working on the assumption there is _always_ at least one 'sent' annual bill run for a region. This is the case in our live service but not in our test setup.

We've made changes to our [acceptance test fixture setup](https://github.com/DEFRA/water-abstraction-service/pull/2484) to ensure we always create an annual as part of our SROC test fixture. These changes cover any other tweaks needed to get the existing tests working.

---

Notes on the changes

- `cypress/e2e/internal/billing/supplementary/cancel-existing.cy.js`

It was 'cheating' before by using a two-part tariff test fixture that would result in empty bill runs. We have only added the annual bill run entry to our SROC data fixture so we needed to switch it to use that and update the assertions accordingly.

- `internal/billing/supplementary/journey.cy.js

Having moved into 2024-25 we've hit a point where part of the PRESROC charge version is no longer included in the bill run.

It starts on Jan 1 2019. When the billing period was 2023-24 five years before that year it would include the billing period 2018-19. This means there would be 4 supplementary bills with the first just being for 2019-01-01 to 2019-03-31.

Now the billing period is 2024-25 the period 2018-2019 has dropped off. This means 1 less bill which affects the bill run total.

- `internal/billing/supplementary/replace-cv-current-year-with-changes.cy.js`

Moving into a new billing period means the number of billable days being calculated is different which affects the values being asserted.

- `internal/billing/supplementary/change-billing-account-current-year.cy.js`

This corrects a faulty assumption in the test, put there no doubt because of a copy and paste. Once we add the new charge version for the current year and then create the second SRO supplementary, with no other changes we should only see 2 bills; the credit for the previous one and the debit for the new account.

But we were using the number returned from our logic which works out how many bills we expect to see based on the number of billing periods since 2022. It worked when running the tests in 2023-24 because the result was also 2. Now we are in 2024-25 it's exposed the faulty logic.

- Fix broken annual billing

We had overlooked that the annual billing tests were using the `sroc-billing-data` fixture as well. Adding an annual bill run to that fixture caused the annual tests to fail because of the existing bill run.

We realised that we could use the 'previous' fixture we created as part of these changes both for our new supplementary test and the annual tests. This saves a bunch of duplication in the integration sets. It still creates a bill run but it is for the previous year. If anything it's a more robust way of checking that bill runs in previous billing periods shouldn't prevent a new one from being created.

- Add new test for supplementary financial year bump

This new test gives us a means to check that the supplementary end-year bump feature is working. It uses a new SROC fixture that is the same as our normal one only the annual bill run created is for the previous billing period.

This means we can confirm when we create an SROC supplementary that its financial year end is also for the previous billing period.