Update tests for new bill runs page
https://eaflood.atlassian.net/browse/WATER-4445

In [Migrate view bill runs page from legacy UI](https://github.com/DEFRA/water-abstraction-system/pull/925) we were able to implement a new version of the bill runs page!

But that means we need to update any tests that interact with it to use the new selectors we have made available.