Update billing tests
Use new data-test attributes wherever possible. Also, realised we kept including asserts inside getters to local variables, for example

```javascript
  cy.get('@formattedCurrentDate').then((formattedCurrentDate) => {
    cy.get('tr:nth-child(1)')
      .should('contain.text', formattedCurrentDate)
      .and('contain.text', 'Test Region')
      .and('contain.text', 'Annual')
  })
```

We only needed to assert the date value inside the `cy.get('@formattedCurrentDate')` call. But it makes it look like you have to do them all for the test to work for the uninitiated. So, where we have spotted stuff like this we have also fixed it up.