No returns required journey acceptance tests
https://eaflood.atlassian.net/browse/WATER-4479

As the return requirements journey is now nearing completion, we are starting to
create automated acceptance tests for the returns setup functionality.

This also a chance for team members with no experience with [cypress](https://docs.cypress.io/guides/overview/why-cypress)
to get to grips with the framework and the [water-abstraction-acceptance-tests](https://github.com/DEFRA/water-abstraction-acceptance-tests).

This PR is primarily focused on creating the acceptance tests for the no-returns required journey.