Scenario one for the two-part tariff review pages
https://eaflood.atlassian.net/browse/WATER-4452

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the difference scenarios we mapped out at the begging of starting this work.

This PR is focusing on scenario one. This scenario is the simplest `happy-path` test scenarios we can think of.
The test is setting up a new two-part tariff bill run with the data from scenario one, and then checking the review pages are displayed as expected.