Scenario two for the two-part tariff review pages
As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the different scenarios we mapped out at the beginning of this work.

This PR focuses on scenario two of our two-part tariff scenario spreadsheet which we used during the creation of our engine.
This scenario involves having non two-part tariff elements and returns. These should be completely ignored by our engine.