Add cancel a returns required acceptance test
https://eaflood.atlassian.net/browse/WATER-4486

As the work for the return requirements journey is now nearing completion, we are starting to create
automated acceptance tests for the returns setup functionality.

This PR is focused on creating the acceptance tests for cancelling a returns requirement. Since both
the no-returns and returns required journeys both use the same functionality for cancelling a requirement,
it makes sense to test under the no-returns required journey as it is the shorter of the two routes.