Scenario one for the two-part tariff review pages (#92)
https://eaflood.atlassian.net/browse/WATER-4452

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the different scenarios we mapped out at the beginning of this work.

This PR focuses on two testing scenarios. The first is an empty bill run. This is testing the bill run being created, it is empty and then deleting the bill run. The second is called 'Scenario one' as per the scenario spreadsheet we used during the creation of our two-part tariff engine. This scenario is the simplest happy-path test scenario we can think of.
The test is setting up a new two-part tariff bill run with the data from 'scenario one', and then checking the review pages are displayed as expected.