Altering two-part tariff scenario one fixtures (#95)
During the testing of scenario one, it was noticed that the fixtures file was set up incorrectly, meaning the QA team could not run the tests. This is due to an ID being used that was only available on my local machine and therefore not in the QA team database. This should not have been a hardcoded value, but rather a lookup for the correct ID.

This PR is changing the ID from a hardcoded value to a lookup value instead.