Returns required journey acceptance tests (#91)
* Returns required journey acceptance tests

https://eaflood.atlassian.net/browse/WATER-4485

As the return requirements journey is now nearing completion, we are starting to create automated acceptance tests for the returns setup functionality.

This PR is primarily focused on creating the acceptance tests for the returns required journey. It includes tests for changing selected options throughout the journey as well as adding and removing a requirement.