Updating testing scenario one for two-part tariff review pages
https://eaflood.atlassian.net/browse/WATER-4452

Scenario one for the two-part tariff review screens is our happy path scenario. This testing scenario was written whilst live testing was happening in our pre production environment with the billing and data team. During their testing of the two-part tariff engine and review screens they requested some changes to make it more user friendly. This PR is updating the acceptance test for scenario one to cater for the changes to the review page that have been made.