Fix two-part tariff scenario one
When updating test scenario one to account for new engine changes we made, I updated had updated a test attribute on the review page. However I forgot to change the attribute on the acceptance test which now means it fails. This PR is just updating the attribute tag to make it correct.