Scenario four for the two-part tariff review pages
https://eaflood.atlassian.net/browse/WATER-4498

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario four of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having a charge reference with an aggregate. This should be flagged by the engine.