Add cancel a return requirement acceptance test (#94)
* Add cancel a returns required acceptance test

https://eaflood.atlassian.net/browse/WATER-4486

As the work for the return requirements journey is now nearing completion, we are starting to create automated acceptance tests for the returns setup functionality.

This PR is focused on creating the acceptance tests for cancelling a returns requirement. Both the no-returns and returns required journeys both use the same functionality for cancelling a requirement.  For this PR, the tests will follow the returns required journey to confirm the cancel requirement functionality.