Scenario five for the two-part tariff review pages (#102)
https://eaflood.atlassian.net/browse/WATER-4499

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Five of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having returns that overlap on the charge dates. This should be flagged as an issue on the engine.