Scenario Eight for the two-part tariff review pages (#105)
https://eaflood.atlassian.net/browse/WATER-4502

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Eight of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having nil returns. This should not be flagged as an issue and should not allocate any return quantities to the charge element since it is a nil return.