Fixing broken two-part tariff acceptance tests
https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/106

While writing the acceptance test for scenario nine in our two-part tariff testing scenarios I had to update the tags on the licence review page. This was because this is the first scenario where we have multiple charge versions, charge references and charge elements and it became clear that the tags weren't handling the indexes of these correctly. Once the tags where updated the rest of the acceptance tests broke, meaning this PR is me fixing them!