Scenario Nine for the two-part tariff review pages (#106)
https://eaflood.atlassian.net/browse/WATER-4503

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Nine of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having a licence without return submissions. This should flag multiple issues of 'No returns received' and 'Some returns not received'. Both charge elements should be fully allocated.