Scenario Ten for the two-part tariff review pages (#108)
https://eaflood.atlassian.net/browse/WATER-4504

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Ten of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having a licence with no matching returns. This should flag the 'Unable to match return' issue and nothing should be allocated.