Scenario Fourteen for the two-part tariff review pages (#114)
https://eaflood.atlassian.net/browse/WATER-4509

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Fourteen of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having a licence with a return that matches two charge elements. This should raise no issues on the engine. This scenario demonstrates how the charge elements are allocated in order of the highest authorised volume first. It also demonstrates how a line on a return can be split over multiple charge elements during the allocation process.