Scenario Thirteen for the two-part tariff review pages (#113)
https://eaflood.atlassian.net/browse/WATER-4508

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Thirteen of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having a licence with returns that don't match a charge element. This should flag the 'Unable to match return' issue and the return should not be allocated. The return should be in its unmatched section.