Scenario Twelve for the two-part tariff review pages (#112)
https://eaflood.atlassian.net/browse/WATER-4507

As the work for the two-part tariff engine and review pages is nearing completion, we are starting to create automated acceptance tests for the scenarios we mapped out at the beginning of this work. This PR focuses on Scenario Twelve of our two-part tariff scenario spreadsheet which we used during the creation of our engine. This scenario involves having a licence with returns split over charge references. This should flag the 'Split over charge reference' issue and allocate the return to the highest subsidence charge first.