Fix the two-part tariff scenario acceptance tests
During a change to the licence review page for the two-part tariff review process, all links that direct a user to view the licence were changed along with their data test attributes. This change broke all the acceptance tests, as each test is looking for a specific tag to check that the links exist. This PR is updating the acceptance tests to look for the new tags.