Two-part tariff test scenarios housekeeping
https://eaflood.atlassian.net/browse/WATER-4450

All our acceptance tests have been written for the two-part tariff engine and licence review pages. Now that work has been done to keep things clean and documented this PR is some house keeping to add a README file and change the folder layout of the tests. The two-part tariff scenarios are quite complicated so hopefully this will make it easier to understand what they are doing when we look back at it.