Fix two-part tariff scenario one acceptance test (#115)
https://eaflood.atlassian.net/browse/WATER-4450

During a change to the licence review page for the two-part tariff review process, all links that direct a user to view the licence were changed along with their data test attributes. This change broke scenario one of the acceptance tests, as the test is looking for a specific tag to check that the links exist. This PR is updating the acceptance test to look for the new tags.