Fix non-working billing tests (#117)
https://eaflood.atlassian.net/browse/WATER-3981

In [Switch to new test data loader for SROC billing](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/90) we started the move to a new acceptance test data loader to replace the slow and flaky legacy one. 😍🫶 

We _swear_ all the tests were passing before we merged. But low and behold a number are failing. When we dug in we found its because there are tests referring to the old licence reference. Clearly, we must have imagined the passing tests! 🤦 

But we also found that those that rely on adding a new charge version were also broken because our fixture wasn't creating data that the legacy charge version setup journey requires.

We added the ability to create the data needed in [Add company address model and relationships](https://github.com/DEFRA/water-abstraction-system/pull/1140).

This change updates our SROC billing fixture and all the references to the old licence reference.