Fix 2PT fixture data
https://eaflood.atlassian.net/browse/WATER-4450

We've been adding a series of acceptance tests to ensure we have regression coverage for any breaking changes to our two-part tariff matching and allocating engine.

All do work and test the things we need to cover. But we have found that we are getting intermittent failures.

The issue seems to come down to sorting; how the versions, references, elements etc are order on the page can change between test runs because we still haven't quite nailed down how to order them.

When we dug into the fixtures that are generating the data though, we found in some cases they were not realistic. For example, a licence having two charge references that are exactly the same. In reality they might have the same charge category, but we would then expect them to have different start and end dates. Or if the same start and end dates, then the charge categories of the references would be different.

This change reviews the 2PT fixture data and updates them where necessary to reflect how the data should be structured.