Restructure scenario one fixture data
To make it easier to see the complete picture for each scenario we'll move to having one fixture file per scenario.

But they do all use the same licence to to save some duplication we move those elements to a reusable fixture. We also expand the data in the file to include elements to make it possible to search for and view the licence. This makes it much easier to debug issues.