Restructure scenario 9
One of the key changes we've made in this scenario is to ditch the second charge version. If a licence has 2 charge versions, it is more likely they will be for the same purpose and therefore charge reference but something else has changed, for example, the billing account, the authorised amount, or the agreements against it.

More likely that a licence has a charge version covering 2 different purposes each with different authorised amounts and details. So, we've taken the second charge reference and element and linked them to the existing charge version, then updated the test accordingly.