Restructure scenario 11
The test data had simply duplicated the charge references which wasn't realistic and could cause issues with the test because the ordering again can not be predetermined.

So, we've made the data more realistic by having one charge reference covering one part of the year and the other the second. We've then ensured the return details line up with this so we only get `Return split over charge references` reported as an issue.