Update acceptance tests to new view licence page
https://eaflood.atlassian.net/browse/WATER-4555

With us creating a new view licence page and making it the default, the ‘elements’ in the page will have changed.
Acceptance tests rely on finding and interacting with the elements of a page, so this means some of our existing tests will fail.

This PR will focus on finding the tests affected by the changes to the elements and updating them to the updated elements.