Fix failing requirement due to new reasons added (#121)
* Fix failing requirement due to new reasons added

https://github.com/DEFRA/water-abstraction-system/pull/1169

During the development stage of the return requirements journey, we seem to have missed out on including
two options from the reason page. The new options have now been added to the journey but as a result, some of
our acceptance tests are now failing.

This PR is focused on fixing the acceptance tests failing due to the two new options added to the reason page.