Update two-part tariff acceptance tests (#122)
https://eaflood.atlassian.net/browse/WATER-4513
https://eaflood.atlassian.net/browse/WATER-4548
https://eaflood.atlassian.net/browse/WATER-4549

While testing the two-part tariff review pages, the billing and data team testing the engine wanted a few changes. These were:

To add a warning to the licence review page if the total sum of the charge elements allocated quantity exceeds that of its charge reference authorised quantity
Add the returns abstraction period to the licence review page
Make the total for any two-part tariff bill run in review blank on the bill runs page
This PR is to update the acceptance tests with these changes.