Replace two part tariff test data (#125)
https://eaflood.atlassian.net/browse/WATER-4523

As part of our work to replace the legacy code, we’ve tackled the complexities and failures of the old test data loader that hindered our acceptance tests. This PR switches the two-part tariff acceptance tests to use the new fixture file in the acceptance tests repo, eliminating the dependency on the service repo's test data.