Migrate bulk return fixture data
https://eaflood.atlassian.net/browse/WATER-4523

As part of our work to replace the legacy code, we’ve tackled the complexities and failures of the old test data loader that hindered our acceptance tests. This PR switches the remaining acceptance tests to use the new fixture file in the acceptance tests repo, eliminating the dependency on the service repo's test data.

We are skipping the bulk acceptance test for now, as it currently fails. More details about this test can be found [here](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/2).

Since fixing this test remains on our to-do list and the data being loaded from the service repository is scheduled for deletion (now that it has been fully migrated), this PR migrates the data file from the service repository to the acceptance test repository. This allows us to delete the data from the service repository while still retaining it for future use in testing. Fixing the test is still on the to-do list.