Delete bulk-upload acceptance test (#127)
https://eaflood.atlassian.net/browse/WATER-4523

As part of our work to replace the legacy code, we’ve tackled the complexities and failures of the old test data loader that hindered our acceptance tests. This PR switches the remaining acceptance tests to use the new fixture file in the acceptance tests repo, eliminating the dependency on the service repo's test data.
We are skipping the bulk acceptance test for now, as it currently fails. More details about this test can be found here.

Since we have been skipping this test since the team started and the data being loaded from the service repository is scheduled for deletion (now that it has been fully migrated), we have decided to delete the test file instead of migrating the data over.