Replace loaded users with seeded ones (#128)
#119

During the last acceptance test housekeeping PR, we migrated all the fixtures data over from the service repo into our acceptance tests repo. Whilst this work was happening, we also started work on seeding certain db tables that had fixed data in them. We have seeded test users already in place, so this PR is to remove the users being loaded in through a fixture file, in favour of our seed. This should increase the speed of the acceptance tests run, as there will be no need to insert and tear down the same users for every test.