Replace loaded region with seeded ones
https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/119

During the last acceptance test housekeeping PR, we migrated all the fixtures data over from the service repo into our acceptance tests repo. Whilst this work was happening, we also started work on seeding certain db tables that had fixed data in them. We have now seeded the (region)[https://github.com/DEFRA/water-abstraction-system/pull/1208], so this PR is to remove the regions being loaded in through a fixture file, in favour of our seed. This should increase the speed of the acceptance tests run, as there will be no need to insert and tear down the same region for every test.