Fix Return Requirements fixtures file (#130)
After running the acceptance tests multiple times our QA team discovered that the return requirements tests were randomly failing on the data load.

As we know nothing is ever random...

The return requirements tests all pull in the same fixtures file returns-requirements.json. The load service was failing with an error duplicate key value violates unique constraint "uindx_licence_version_purposes_external_id". On the fixtures file for licenceVersionPurposes, the externalId wasn't hard-coded data meaning it was using our helper in the system repo. The helper in the system repo is randomly generating the externalId each time.

My guess is that because the acceptance test is using our local db, which in the licenceVersionPurposes table has thousands of records, sometimes the randomly generated externalId clashed with a real record we have and therefore failed to load. With this change, we have hardcoded the ID to avoid this clash. After spamming the acceptance tests it has passed every time.