Add external user change email test
https://eaflood.atlassian.net/browse/WATER-4632

We received a report that the external apps change email journey was broken. We were quickly able to confirm this was the case.

We've been able to locate the problem and fix it. But such a critical feature shouldn't have been allowed to sit broken for so long.

To ensure this (till such time as we can replace the legacy version!) we're adding this test so we have some coverage of the feature.