Add external user change email test (#131)
https://eaflood.atlassian.net/browse/WATER-4632

We received a report that the external app's change email journey was broken. We quickly confirmed this.

We've been able to locate and fix the problem, but such a critical feature shouldn't have been allowed to be broken for so long.

To ensure this (until we can replace the legacy version!), we're adding this test so we have some coverage of the feature.