Update fixtures to lookup region
https://github.com/DEFRA/water-abstraction-team/issues/124

As part of our ongoing efforts to remove the dependency on cleaning the DB before a test, we introduced seeders for some of the reference data we depend on. That means, for example, you cannot add a region using our test helpers and the loaders any more.

In [Replace loaded region with seeded ones](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/129) we updated our fixtures because of this. Other reference records are already using lookups instead of hard coding seeded ID's and on reflection, we should have done the same for region.

We cannot 'guarantee' that the test region will have _definitely_ been seeded with the ID we provided. For reasons out of our control it might have a different one. Plus there is the concept of consistency; why have a hard coded ID when we have lookups for others.

So, this updates any fixture that is reliant on creating a record linked to the test region to use a lookup instead of the hard coded ID.