Update all scenarios to use new points solution
https://eaflood.atlassian.net/browse/WATER-4600
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

The new return version setup journey requires at least one point be selected for each return requirement in the return version. The points to select from come from those currently linked to the licence. Our initial version of the solution copied what the legacy code was doing when it needed points, extracting them from the JSONB 'mega-blob' in `permit.licence` table. There simply was no other source.

We then hit a problem in that [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) wasn't populating this information once a licence was 'ended'. Users need to be able to correct and update the return version history even for ended licences. We resolved to start importing NALD point information into new tables, along with details of which return requirements and licence version purposes were linked to them. It also includes the water sources for each point.

With this new data source in place we've updated all functionality ([water-abstraction-system only!](https://github.com/DEFRA/water-abstraction-system)) that depends on points to use the new source.

Of course, this breaks a number of the acceptance tests because they are no longer populating the right tables. This change updates the fixtures to get them working again!