Fix external login journey for blank environment
We hit an issue recently where the `cypress/e2e/external/login/journey.cy.js` test worked for the devs but not one of the QA's.

When we looked into it we found that the assertion was based on the external user logging in already having licences linked to their account.

Before we migrated the tests this would have been the case because the user was created as part of a large import of data referred to as 'barebones'. (psst! It wasn't that bare!)

Once migrated and after we switched to our new data seeder we dropped the loading of a fixture for this test. All non-prod environments now seed the same users, so for a simple login test there should be no need to load a fixture.

It was working for most users because at some point, we've linked a licence to the seeded external user as part of other testing/investigations.

But the QA this test was failing for had never done that. So, the element the test was asserting against never shows.

We can fix the test by instead asserting against an element that appears whether a the user has linked licences or not.