Update 2PT no returns matched scenarios
https://eaflood.atlassian.net/browse/WATER-4684

After user testing the new two-part tariff matching and allocating engine we've built (part of two-part tariff billing) we were asked to make a tweak.

When a charge reference has no matching returns, we were defaulting the allocation to 0. So, we have two review scenarios that ensure this.

We've been asked to change this to instead allocate the authorised volume on the charge reference. Having made the change, it means these scenarios are now failing.

This change updates them to match the new behaviour.