Add Supplementary Billing Flags tests
https://eaflood.atlassian.net/browse/WATER-4746

As part of the ongoing effort to clean up and refactor the code surrounding supplementary billing flags, this PR is focused on adding in acceptance tests that will test the various routes that require a licence to be flagged.

There are multiple ways in which a licence can be flagged for supplementary billing. These are:

* Deleting a licence from workflow
* Cancelling a charge version (that is waiting to be approved)
* Approving a charge version

* Removing a licence from a bill run (pre sroc two-part tariff, annual, two-part tariff annual and supplementary)
* Editing a return
* Using the recalculate bills link
* A licence ending
* Setting up a charging agreement
* Deleting a charging agreement
* Ending a charging agreement

Each of these various ways sets its own flags for supplementary billing. With the updated code changes they now all hit a ne endpoint in system that determines which flags need to be set, but each mechanism does not set every flag. Take the setting up a licence agreement as an example. This is legacy code that we have migrated to keep it updated and working, but this route only flags for pre-sroc and not sroc and two-part tariff.

The acceptance tests in this PR aims to cover all the various routes that flagging can happen.