Update abstraction alert tests after changes
https://eaflood.atlassian.net/browse/WATER-4736

We have migrated the view monitoring stations page from the legacy code to [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system). Doing so we have updated the page design and used current GDS design components. However, these changes have caused some of the tests to break.

This change updates them to be able to locate the updated elements on the page.