Update 2PT review tests to match actual titles
https://eaflood.atlassian.net/browse/WATER-4201

> Part of the work to support two-part tariff bill runs

As part of updating the service to support two-part tariff supplementary bill runs we needed to update the review pages to ensure the title reflected the two types of two-part tariff bill run.

That's when we spotted that the existing functionality is not using our bill run title helpers which ensure titles are displayed consistently and correctly. We updated [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/1562) to use the new helpers. but this has meant the expected titles no longer quite match what we generate.

This change fixes the tests.