add historic returns acceptance test sceanorios (#150)
https://eaflood.atlassian.net/browse/WATER-4892

This PR adds two test scenarios for historic returns. One is to check a winter and all year only change and the other is a test for both summer and winter and all year return cycles.

This PR adds some of the logic for determining the due dates from the system repo so that when the tests are ran they always use current return cycle details.