Fix broken billing tests caused by year change
We've gone to run our acceptance tests and found some have failed. All are connected to the switch into the next financial year, and these changes are for failing supplementary tests.

As we move away from 2022, the number of bills expected in supplementary bills runs will increase, whilst PRE-SROC will reduce. In some places we're handling this with calculated assertions, but it looks like we missed some.

This change fixes those omissions.