Replace numberOfSrocBillingPeriods with generic
The reason the tests for the presroc assertions failed is because they are hard code. To avoid that in future years we need to make the count dynamic like we do for SROC.

We then realised that if we know the SROC count, working out the PRESROC becomes trivial. So, why not combine the functionality into a single function?